+++ src/index.js
@@ -10,3 +10,4 @@ export {
   default as useIsOffline,
   IsOfflineContextProvider,
 } from './hooks/useIsOffline'
+export { default as useLogin } from './hooks/useLogin'