+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- `useLogin()` hook to help with implementing a custom login screen
+
 ### Fixed
 
 - Scroll service targetting class that wasn't present on DOM